# apps/api/swagger/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/projects/{projectId}/batches": {
      "post": {
        "operationId": "BatchesController_create",
        "parameters": [
          {
            "name": "projectId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateBatchDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BatchDto"
                }
              }
            }
          }
        },
        "summary": "Create a new batch for script generation",
        "tags": [
          "Batches"
        ]
      },
      "get": {
        "operationId": "BatchesController_findAllByProject",
        "parameters": [
          {
            "name": "projectId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/BatchDto"
                  }
                }
              }
            }
          }
        },
        "summary": "List all batches for a project",
        "tags": [
          "Batches"
        ]
      }
    },
    "/batches/{id}": {
      "get": {
        "operationId": "BatchesController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BatchDto"
                }
              }
            }
          }
        },
        "summary": "Get a batch by ID",
        "tags": [
          "Batches"
        ]
      }
    },
    "/batches/{id}/scripts": {
      "get": {
        "operationId": "BatchesController_getScripts",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ScriptDto"
                  }
                }
              }
            }
          }
        },
        "summary": "Get all scripts for a batch",
        "tags": [
          "Batches"
        ]
      }
    },
    "/scripts/{id}/regenerate": {
      "post": {
        "operationId": "BatchesController_regenerateScript",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegenerateDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ScriptDto"
                }
              }
            }
          }
        },
        "summary": "Regenerate a script with new instructions",
        "tags": [
          "Batches"
        ]
      }
    },
...
    "/admin/requests": {
      "get": {
        "operationId": "AdminController_getRequests",
        "parameters": [
          {
            "name": "status",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Admin"
        ]
      }
    },
    "/admin/requests/{id}/approve": {
      "post": {
        "operationId": "AdminController_approveRequest",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Admin"
        ]
      }
    },
    "/admin/requests/{id}/reject": {
      "post": {
        "operationId": "AdminController_rejectRequest",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Admin"
        ]
      }
    },
    "/admin/requests/{id}": {
      "delete": {
        "operationId": "AdminController_deleteRequest",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Admin"
        ]
      }
    },
...
    "/admin/users/{id}": {
      "delete": {
        "operationId": "AdminController_deleteUser",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Admin"
        ]
      },
      "get": {
        "operationId": "AdminController_getUserDetail",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminUserDetailDto"
                }
              }
            }
          }
        },
        "summary": "Get user details with credits",
        "tags": [
          "Admin"
        ]
      }
    },
    "/admin/users/{id}/toggle-admin": {
      "post": {
        "operationId": "AdminController_toggleAdmin",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Admin"
        ]
      }
    },
    "/admin/users/{id}/magic-link": {
      "post": {
        "operationId": "AdminController_generateMagicLink",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Admin"
        ]
      }
    },
    "/admin/users/{id}/plan": {
      "patch": {
        "operationId": "AdminController_updateUserPlan",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Admin"
        ]
      }
    },
    "/admin/users/{id}/credits": {
      "post": {
        "operationId": "AdminController_grantCredits",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GrantCreditsDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GrantCreditsResponseDto"
                }
              }
            }
          }
        },
        "summary": "Grant credits to user",
        "tags": [
          "Admin"
        ]
      }
    },
...
    "/admin/queue/jobs": {
      "get": {
        "operationId": "AdminController_getQueueJobs",
        "parameters": [
          {
            "name": "status",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Admin"
        ]
      }
    },
    "/admin/queue/jobs/{id}/retry": {
      "post": {
        "operationId": "AdminController_retryJob",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Admin"
        ]
      }
    },
...
  },
  "info": {
    "title": "Script Factory API",
    "description": "API for UGC Script Factory",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "cookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "access_token"
      }
    },
    "schemas": {
...
      "CreateBatchDto": {
        "type": "object",
        "properties": {
          "requestedCount": {
            "type": "number",
            "minimum": 1,
            "maximum": 200
          },
          "platform": {
            "type": "string",
            "enum": [
              "universal",
              "tiktok",
              "reels",
              "shorts"
            ]
          },
          "angles": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "durations": {
            "type": "array",
            "items": {
              "type": "number"
            }
          },
          "personaIds": {
            "description": "Empty or omitted = all personas",
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "quality": {
            "type": "string",
            "enum": [
              "standard",
              "premium"
            ],
            "default": "standard"
          }
        },
        "required": [
          "requestedCount",
          "platform",
          "angles",
          "durations"
        ]
      },
      "BatchScriptsCountDto": {
        "type": "object",
        "properties": {
          "scripts": {
            "type": "number"
          }
        },
        "required": [
          "scripts"
        ]
      },
      "BatchDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "status": {
            "type": "string",
            "enum": [
              "pending",
              "processing",
              "completed",
              "failed"
            ]
          },
          "requestedCount": {
            "type": "number"
          },
          "platform": {
            "type": "string"
          },
          "angles": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "durations": {
            "type": "array",
            "items": {
              "type": "number"
            }
          },
          "personaIds": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "quality": {
            "type": "string",
            "enum": [
              "standard",
              "premium"
            ]
          },
          "pdfUrl": {
            "type": "string"
          },
          "csvUrl": {
            "type": "string"
          },
          "errorMessage": {
            "type": "string"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string"
          },
          "scriptsCount": {
            "type": "number"
          },
          "completedCount": {
            "type": "number"
          },
          "progress": {
            "type": "number"
          },
          "_count": {
            "$ref": "#/components/schemas/BatchScriptsCountDto"
          }
        },
        "required": [
          "id",
          "status",
          "requestedCount",
          "platform",
          "angles",
          "durations",
          "personaIds",
          "quality",
          "createdAt",
          "updatedAt"
        ]
      },
      "StoryboardStepDto": {
        "type": "object",
        "properties": {
          "t": {
            "type": "string"
          },
          "shot": {
            "type": "string"
          },
          "onScreen": {
            "type": "string"
          },
          "spoken": {
            "type": "string"
          },
          "broll": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        },
        "required": [
          "t",
          "shot",
          "onScreen",
          "spoken"
        ]
      },
      "ScriptDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "status": {
            "type": "string",
            "enum": [
              "pending",
              "generating",
              "completed",
              "failed"
            ]
          },
          "angle": {
            "type": "string"
          },
          "duration": {
            "type": "number"
          },
          "hook": {
            "type": "string"
          },
          "storyboard": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/StoryboardStepDto"
            }
          },
          "ctaVariants": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "filmingChecklist": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "warnings": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "score": {
            "type": "number"
          },
          "errorMessage": {
            "type": "string"
          },
          "parentScriptId": {
            "type": "string",
            "description": "ID of the parent script if this is a regenerated version"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string"
          }
        },
        "required": [
          "id",
          "status",
          "angle",
          "duration",
          "ctaVariants",
          "filmingChecklist",
          "warnings",
          "createdAt",
          "updatedAt"
        ]
      },
      "RegenerateDto": {
        "type": "object",
        "properties": {
          "instruction": {
            "type": "string"
          }
        },
        "required": [
          "instruction"
        ]
      },
...
      "AdminUserDetailDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "isAdmin": {
            "type": "boolean"
          },
          "plan": {
            "type": "string"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "subscriptionStatus": {
            "type": "string"
          },
          "subscriptionEndsAt": {
            "format": "date-time",
            "type": "string"
          },
          "_count": {
            "type": "object"
          },
          "creditBalances": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "recentTransactions": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        },
        "required": [
          "id",
          "email",
          "isAdmin",
          "plan",
          "createdAt",
          "_count",
          "creditBalances",
          "recentTransactions"
        ]
      },
      "GrantCreditsDto": {
        "type": "object",
        "properties": {
          "creditType": {
            "type": "string",
            "enum": [
              "free",
              "subscription",
              "pack"
            ]
          },
          "amount": {
            "type": "number"
          },
          "description": {
            "type": "string"
          }
        },
        "required": [
          "creditType",
          "amount"
        ]
      },
      "GrantCreditsResponseDto": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean"
          },
          "newBalance": {
            "type": "number"
          }
        },
        "required": [
          "success",
          "newBalance"
        ]
      },
...
    }
  }
}